feat: ui update and polar product chosen
