+++ apps/web/src/lib/polar.ts
@@ -1,7 +1,13 @@
-import { Checkout } from '@polar-sh/nextjs';
+// import { Checkout } from '@polar-sh/nextjs';
+// export const GET = Checkout({
+//   accessToken: process.env.POLAR_ACCESS_TOKEN || '',
+//   successUrl: '/success',
+//   server: 'sandbox',
+// });
+import { Polar } from '@polar-sh/sdk';
 
-export const GET = Checkout({
-  accessToken: process.env.POLAR_ACCESS_TOKEN || '',
-  successUrl: '/success',
-  server: 'sandbox',
+console.log(process.env.NEXT_PUBLIC_POLAR_ACCESS_TOKEN);
+export const api = new Polar({
+  accessToken: process.env.NEXT_PUBLIC_POLAR_ACCESS_TOKEN || '',
+  server: 'production',
 });
